Simplify and improve exception/signal handling
Previously, it was possible to have some S3cmds execute after the temp
file with the configuration had been deleted, as the parent process
would abort before the children. This "finally" block is more robust
to other kinds of raised exceptions and gets cleanup to happen at the
right time.

Signed-off-by: Daniel Farina <daniel@heroku.com>