More signal safety: subprocess s3cmd holding
The goal here is to propagate SIGINT to a child process (this may be
done automatically, but we're making doubly sure in case of obscure
possibilities in the subprocess library/Python interpreter) and then
wait for it to return before blowing away the temporary config file.

Otherwise, the order of events may be wrong: the file will get blown
away while the process is starting up, which will then be confused by
the lack of said file.

Signed-off-by: Daniel Farina <daniel@heroku.com>