Add file structure versioning
This in the form of a number that can be used to discriminate what
version of wal-e was used to prepare the backup.  For example,
versions of wal-e previous to this may have stored the sentinel file
compressed as .lzo, or uncompressed.  Future versions may pack small
files, or not.  And so on.

Signed-off-by: Daniel Farina <daniel@heroku.com>