Add note about blocking on sending WAL even upload didn't succeed
Unfortunately there's no particularly sound way out of this: when
calling pg_stop_backup via psql, there is no way to know that the
function invocation has made its way to the backend, and it insists on
blocking until WAL segments are uploaded, even if the base backup did
not succeed.

This probably could be addressed by sending a query cancel with
psycopg2 or any other lower level driver, or if Postgres implemented a
pg_cancel_backup or similar.  Regardless, no such thing exists for
8.3, the minimum compatibility target.

Signed-off-by: Daniel Farina <daniel@heroku.com>