Make pool_size a required argument
Set the default from argparse.  This may not be very good api
design (pool size should probably be keyworded, with a
reasonable default), but this reduces the amount of noise in
*this* program.

The urge to do this was brought on by accidentally setting
pool_size to 1 from argparse, and then not noticing for some
time -- the defaults in the keywords were never used.

Signed-off-by: Daniel Farina <daniel@heroku.com>