Timezone pedantry
This was way too difficult.  To the best of my knowledge, the
timestamp used for the backup name is not used anywhere, but when
fumbling through the program in my mind I realized this oversight.

Credit to kr@heroku.com for engaging my timezone pedantry a few weeks
previously, where I argued that integer epoch times as an external
representation are a step backward, and one should just use timezone
labeling instead.  I am now taking my own good advice.

Signed-off-by: Daniel Farina <daniel@heroku.com>