Begin to have some semblance of user error reporting
This is accomplished through a new exception type, with msg, detail,
and hint components, similar to PostgreSQL (stolen outright because it
seemed as good/better than many guidelines, thorough, and not
something I had to write myself).

These errors are written to a logger.  One day environment variables
or configuration may be used to bend the logger to someone's will.

Conflated in this commit: a somewhat more elaborate import/multi-file
structure that occurred around the same time.

Signed-off-by: Daniel Farina <daniel@heroku.com>