Improve file-reading pipeline
Change the order -- StreamPadFileObj now comes before mbuffer -- and
handle (that is to say, ignore) unlinked file race conditions.

This is in some ways cleaner (avoiding killing mbuffer with SIGPIPE)
but in particular lets Python handle the ENOENT in event of an
unlinked file.

The amount of error handling nesting involved was getting unwieldy, so
a new symbol to address that code has been introduced.

Signed-off-by: Daniel Farina <daniel@heroku.com>