Remove bogus attempts at robustness from sentinel upload
This code dates back to some of the earliest versions of wal-e, back
when it relied on s3cmd, and was a pitiful attempt basic robustness.
So stop pretending its robust, update the caveats, and salt the earth
with warnings and promises to make things better and particular ways
to do that.

Signed-off-by: Daniel Farina <drfarina@acm.org>