Emit correct error message when there is a problem with psql
Mechanism: only EnvironmentErrors are being handled nowadays to
trigger the error-message-accumulation and reporting logic, so fake up
an EnvironmentError as to let normal error handling take place here.

This is definitely a kludge, but it doesn't seem worth it perform the
obvious alternative option, which is making a new exception type to
treat the same problem.

Signed-off-by: Daniel Farina <daniel@heroku.com>