Change the logging format: move time and pid to the structured fields
This was raised by Ed Muller <edward@heroku.com> as something that
makes the structured format much more useful.  Unwilling to duplicate
the time and pid information at both the log header and in the
structured information fields, this patch opts to move time to the
structured area.

However, as a service to human readers, the time and pid fields are
guaranteed to come first.  Previously, all fields were lexically
sorted.

Signed-off-by: Daniel Farina <daniel@heroku.com>