Address CR feedback.
In s3_endpoint_for_uri(), update the hard-coded region dictionary with
any entries in boto's S3 region map if available to maximise
compatibility.

Add a unit test and integration test.  This adds a testing dependency on
unittest2 which is a Python 2.6 backport of the unittest module
improvements in Python 2.7.

Having AWS credentials in your environment along with the
WALE_S3_INTEGRATION_TESTS (any arbitrary value) will trigger inclusion
of the integration tests.