setup.py: downcase WAL-E to wal-e
Continue the general conversion of machine-intended references to
WAL-E to be lower case, much like the executable, package name,
repository, and so on.

It is still handy to have capitalized name (and WAL is an acronym
anyway) because it makes references to WAL-E stand out in human-read
text without quoting or special formatting.

Signed-off-by: Daniel Farina <daniel@heroku.com>