Retire code that was used for debugging
This timeout should be unnecessary because it is expected that the
other greenlets will terminate in a timely manner.  If greenlets
aren't terminating in a timely way, I'd rather figure out what the
cause is than masking it.

Luckily, this only causes more failures than otherwise strictly
necessary, as opposed to a kind of bug that can cause WAL to be
skipped.