Fix exit on 404
Most errors deserve a retry, but a notable exception is a missing
archive file.  This regressed because use of boto storage_uri was
deprecated when implementing multi-region support.

As-is, wal-e will spin in a tight loop when encountering such a
missing archive file without ever notifying Postgres that it is not
currently present.