Fix buckets in S3-Classic/US-Standard that have dots
us-standard buckets that have dots ('.'), but are otherwise subdomain
compatible and could successfully run get_location(), which would
return '' (empty string), which is not seen in the region dictionary,
causing WAL-E to crash.

The fix is to re-write a successful get_location() of '' to
'us-standard'.

A regression test has been added.

This was not exercised by previous testing because all of them
combined other strangeness of bucket-name (like upper-case) that would
cause defaulting to us-standard.  Also, those that use an IAM set up
that does not permit get_location would not see this regression.

Reported by Christophe Pettus