Fix __exit__ in contexts
The return value of __exit__ is used to determine if Python should
re-raise exceptions, False meaning 're-raise'.

The dangerous looking one is the one in StreamPadFileObj, where a
.close() that might somehow yield a non-False-ish return could cause
exceptions to be swallowed.

It doesn't seem likely that'd ever be the case, but the code as
written needlessly relies on that for correct behavior and obscures
the intent.

Backpatch this change to 0.5, too.