Move documentation about WALE_S3_INTEGRATION_TESTS
test_s3_worker is among the oldest of the tests and introduced the
WALE_S3_INTEGRATION_TESTS pattern.  Since then, it has enjoyed use in
other tests and its code has been factored out to another module, but
those tests do not document the convention.

So, instead move the advisory about that to that module and explain it
at a higher level.