Disable bucket name validation in s3_worker
This can cause problems when one is using IAM policies that
restrict actions to a prefix in the bucket.

Also avoids one API call in the common case where the bucket is valid,
although later calls that touch the bucket for the first time must be
ready for an invalid bucket to raise exceptions.