Simplify and adjust logging setup
Notably, logging will be set up and sent to stderr regardless of
keyword arguments to configure (that flexibility was never used
anyway) and does not depend if root handlers have already been
registered or not.

Prompted after seeing some of the gratuitous difficulty in debugging
for some people due to lack of output, attempt to streamline and
simplify the logging setup procedure.

There may be a loss here to how well WAL-E composes with other
applications were it used as a library, but there are no such users
known to date nor have they complained, so adjust the trade-off in
favor of a higher chance of getting diagnostics for WAL-E users and
hackers.