Minimize the amount of treatment of argv in 'cmd'
Remove some defaults and fallbacks about argv that were not used.
Because of the 'console_scripts' behavior of setuptools, 'cmd.main'
must be able to accept zero arguments and is going to read 'sys.argv'
anyway, which is also the default behavior of 'argparse'.

This breaks a use case of calling 'cmd.py' directly as a toplevel
entry point for the sake of testing, but no known parties do this.