Fix dangerous thinkos involving "StandardError"
StandardError is intended to include the built-in exceptions in
Python.

This was pernicious in the 'retries' module, as most errors were of
the form of timeouts, had an errno, and inherited from the built-in
EnvironmentError, yet other failure modes -- like
httplib.IncompleteRead -- would not inherit from StandardError, and
escape the retry.

Conflicts:
	tests/test_s3_deleter.py
	tests/test_wabs_deleter.py
	wal_e/retries.py