Refactor blackbox and how configuration and argv is injected
The goal is to allow the 'config' object to manipulate the argv under
test, rather than assuming the test should know how to construct argv
verbatim.

This reflects that argv can be used to set up some amount of WAL-E's
credential in addition to environment variables.

An upshot is that now that the 'main' procedure requires injection via
named fixture, also remove the magicalness of 'autouse=True'.