Remove any claim of avoiding traversal attack via backup recovery
This code is problematic for legitimate use cases where the user has
set up symlinks of things like tablespaces after resolving the
symlink.

Unfortunately, this code was put in place in attempt to make sound a
previously unsound prior approach (looking for traversals in the
member name, e.g. ".."), as a malicious tar archive could contain both
a symlink and then a non-relative member name to mount a traversal
attack anyway.

And previous to that, in WAL-E v0.6, no attempt to avoid this attack
was done at all, via reliance on tarfile's "extractall" method.  So,
regressing v0.7 (now in alpha) back to the previous state of affairs
is no disaster.

Putting aside the idea of malicious archives, part of this is also
mundane program-"oops" safeguarding, so it'd be nice to get some level
of safety back somehow.