Avoid attempting to open and fsync extracted symlinks
This does not work.  Tar files that have symlinks pointing at a
non-existent file should not cause WAL-E to crash.  And, it doesn't
see possible to open a symlink on most platforms, and particularly on
Linux, and without an open() there can be no fsync().  A moment's
research was expended on the matter, but at this time there is no
known portable way to flush the *contents* of the symbolic link's
text.

The reasons to allow such dead-end symlinks in tarfiles is obscure.
Part of this is backwards compatibility with old archives that allowed
them, presuming they'd later be overwritten by the user, interpreted
within a chroot, or reference a file elsewhere within the same data
directory that is yet to be extracted by another volume being
downloaded.