Fix InstanceProfileProvider class for boto 2.24
"profile_name" is now a parameter that must be supported in
"get_credentials".

Yes, this is exactly the "fragile base class" problem, but let's hope
that the mechanisms there become dormant again for a long stretch
again.  Or, switch to botocore or something like that to avoid the
crazy amount of credential management logic hard to opt-out of in
boto.