Add flush progress detection to BufferedWriter
Prior to this patch, PIPE_BUF_BYTES smaller than 65535[0] would cause
infinite loops.  It so happens that that this constant and the floor
of PIPE_BUF_BYTES being 65535 as defined by
pipebuf._configure_buffer_sizes obscures this problem, but it is too
frail to rely on.

Instead, discontinue attempts to flush if the inner loop makes no
attempts to write out any bytes, as so PIPE_BUF_BYTES can be decoupled
from the flush retention size.

[0]: The constant of how much data should remain in the buffer at a
     maximum after flushing