Fix the search for the backup 'LATEST' in listing subroutines
Previously, the sort was on Python objects without a comparitor, so
the resultant sort order was arbitrary and then the last member of
this arbitrary sorting would be selected as the "latest".

Instead, use the BackupInfo.last_modified field to do the sort.

This regression was introduced by
e90973ae83b56197851caa1c5247fab255b25f40, which converted BackupInfo
from a tuple (which was in an order as to sort lexically, which given
Postgres naming conventions is roughly by date) to a more general
Python object.