Fix broken behavior for finding LATEST backup
It had been sorting the list of all backups, and taking the final one... but it
was just sorting python objects, so the sort order is basically where they
showed up in memory.

The comment suggested it wanted to sort lexically, presumably by name (as if,
in some past version of the code, the BackupInfo objects were just string
names).

However, since the BackupInfo objects have a last_modified date, I went ahead
and sorted by that (from what I could tell, all the types of storage always
populate the last_modified date).