Extend fixture to truncate gevent timeouts
Previously, a monkeypatch driven fixture to shorten timeouts was put in
place to expedite testing of non-terminating cases.

Noticing that tests/test_wal_transfer.py::test_finally_execution was
slow, it seems that the killall/joinall interaction was not affected by
that monkeypatch.  So, truncate the timeout of those operations as well.

Doing this exposed that at various point the gevent-fast-sleep stuff got
copied around several times with very minor modification after all this
time.  Refactor it into one copy, netting a reduction in code.