Switch PEP3143 Daemon Implementation
Previously, there was some complaint that python-daemon is an annoying
dependency to have.  What sealed its fate is that it depends on an
obsolete version of "lockfile", and that makes it an unacceptable pain
to vendor it and its dependencies when necessary.

Instead, vendor an implementation from Stephan Schultchen, as seen at
https://pypi.python.org/pypi/pep3143daemon.  This doesn't require
lockfile and is smaller.  This project is distributed under MIT, and the
version copied is from a modified version that fixes an obscure
difference in how processes are exited:
https://github.com/fdr/pep3143daemon/commit/5c091da7f5a912bc35e349fa4d0249915c233813