Reorganize the README
It was already awkward with the alternate backends that exist today,
and the snarl is getting worse with two prospective new backends:
Google Storage, and POSIX-ish File System.

I take this opportunity to draw more emphasis to more important
information and cut down some information that is not as important.