Streamline testing instructions and add Google
Remove superfluous detail from the README.  There is some information in
there that is obsolete: testing Python 2.6 isn't supported for some time
now and so isn't as important as it once was.  And nobody has submitted
a `unittest`-style test for a long time.

Expend some of that gained simplicity to add Google integration test
instructions.