Make storage backends optional
Unfortunately, upstream drivers have been found wanting: they are not
reliable to install.  Most of the trouble is with OpenStack, and GCS
has not been reliable either.

I have long resisted this as I thought it was silly to have the
simplest installation of WAL-E be useless, but the unreliability of
all the dependent drivers has forced my hand.  The last straw was
Google.