Fix boto-importing issues seen on Travis
Though `boto` is compatible with Python 3.4 and Python 3.5, Travis has
a plugin installed that is only Python2 compatible.  As a kludge,
remove the plugin when running on Travis.

This is worrisome  in general because a regular  `WAL-E` install could
run into the same issue, but if the alternative is remain on Python 2,
then there is not much of a choice.