Fix new gevent given coros module deprecation
New versions of gevent no longer have a `coros` module, having renamed
that functionality to `lock`.  There was a deprecation warning for
this, and now it is finally gone for good.

This change has been around long enough that even the oldest supported
version of event in the Tox environment `v1.0-oldest` supports
importing `gevent.lock`, so change this without a concern for
backwards compatibility.